+++ src/数论/费马小定理求逆元.cpp
@@ -1,7 +1,7 @@
 using ll = long long;
 
 // 快速幂
-int power(int a, ll b, int p);
+int power(ll a, ll b, int p);
 // 求逆元
 ll inv(ll x, ll p) {
     return power(x, p - 2, p);
